apps/beeai-sdk/pyproject.toml
[project]
name = "beeai-sdk"
version = "0.1.0"
description = "BeeAI SDK"
readme = "README.md"
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.16",
    "beeai-framework[duckduckgo,wikipedia]>=0.1.31",
    "objprint>=0.3.0",
    "uvicorn>=0.35.0",
    "asyncclick>=8.1.8",
    "sse-starlette>=2.2.1",
    "starlette>=0.47.2",
    "anyio>=4.9.0",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-exporter-otlp-proto-http>=1.35.0",
    "opentelemetry-instrumentation-fastapi>=0.56b0",
    "opentelemetry-sdk>=1.35.0",
    "tenacity>=9.1.2",
    "janus>=2.0.0",
    "uvloop>=0.21.0",
]

[build-system]
requires = ["uv_build>=0.7.12,<0.8.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120
target-version = "py311"
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "UP", # pyupgrade
    "I", # isort
    "B", # bugbear
    "N", # pep8-naming
    "C4", # Comprehensions
    "Q", # Quotes
    "SIM", # Simplify
    "RUF", # Ruff
    "TID", # tidy-imports
    "ASYNC", # async
    # TODO: add "DTZ",   # DatetimeZ
    # TODO: add "ANN",   # annotations
]
lint.ignore = [
    "E501", # line lenght (annyoing)
    "N999", # invalid module name beeai-server (yeah, we use a dash, deal with it)
    "F403", # redundant with Pyright
]
force-exclude = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.12.3",
]
